services/coins-service/tests/requirements.txt
httpx==0.27.*
pytest==8.2.*
pytest-asyncio==0.23.*
pytest-cov==5.0.*
pytest-mock==3.14.*

ConfigArgParse==1.7
SQLAlchemy-Utils==0.41.2
yarl==1.18.3